requirement.txt
absl-py==2.1.0
accelerate==1.6.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
albucore==0.0.24
albumentations==2.0.8
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.0.0
argbind==0.3.9
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
asyncer==0.0.8
attrs==23.1.0
audioread==3.0.1
audiosegment==0.23.0
av==12.1.0
Babel==2.13.1
beautifulsoup4==4.12.2
bidict==0.23.1
bleach==6.1.0
blinker==1.9.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.0
conformer==0.3.2
contourpy==1.2.0
controlnet-aux==0.0.3
cramjam==2.9.1
crcmod==1.7
cryptography==44.0.2
cycler==0.12.1
Cython==3.1.1
dataclasses==0.6
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
descript-audio-codec==1.0.0
descript-audiotools==0.7.2
diffusers==0.33.1
docstring_parser==0.16
easydict==1.13
editdistance==0.8.1
einops==0.8.1
executing==2.0.1
exllamav2==0.2.8
fastapi==0.115.12
fastjsonschema==2.19.0
fastparquet==2024.11.0
ffmpeg==1.4
ffmpy==0.3.2
filelock==3.13.3
filetype==1.2.0
fire==0.7.0
Flask==3.1.0
flatbuffers==24.3.25
flatten-dict==0.4.2
fonttools==4.44.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.3.1
funasr==1.2.6
future==1.0.0
gast==0.6.0
google-pasta==0.2.0
gradio==5.23.0
gradio_client==1.8.0
gradio_vistimeline==1.0.1
graphviz==0.20.3
groovy==0.1.2
grpcio==1.67.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.30.2
humanfriendly==10.0
hydra-core==1.3.2
HyperPyYAML==1.2.2
idna==3.4
imageio==2.37.0
importlib_metadata==8.6.1
importlib_resources==6.4.0
insightface==0.7.3
IOPaint==1.6.0
ipykernel==6.26.0
ipython==8.17.2
isoduration==20.11.0
itsdangerous==2.2.0
jaconv==0.4.0
jamo==0.4.1
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.4.2
json5==0.9.14
jsonpickle==4.0.5
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
julius==0.2.7
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_server==2.10.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.2
kaldiio==2.18.1
keras==3.6.0
kiwisolver==1.4.5
kornia==0.8.1
kornia_rs==0.1.9
langdetect==1.0.9
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.44.0
loguru==0.7.3
lxml==5.2.2
lz4==4.3.3
Markdown==3.7
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
mdurl==0.1.2
midi2audio==0.1.1
mistune==3.0.2
ml-dtypes==0.4.1
modelscope==1.20.1
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
mtcnn==1.0.0
multidict==6.2.0
music21==9.5.0
namex==0.0.8
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.4.2
ninja==1.11.1.4
noisereduce==3.0.3
notebook==7.0.6
notebook_shim==0.2.3
numba==0.61.0
numpy==1.26.2
omegaconf==2.3.0
onnx==1.18.0
onnxruntime==1.21.0
onnxruntime-gpu==1.21.0
openai-whisper==20240930
opencv-contrib-python==4.11.0.86
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.13.0
orjson==3.9.15
oss2==2.19.1
overrides==7.4.0
packaging==23.2
pandas==2.1.2
pandocfilters==1.5.0
parso==0.8.3
patsy==1.0.1
peft==0.7.1
piexif==1.1.3
pillow==11.1.0
pip==25.1.1
platformdirs==4.0.0
pooch==1.8.2
prettytable==3.16.0
prometheus-client==0.18.0
prompt-toolkit==3.0.41
propcache==0.3.1
protobuf==6.31.1
psutil==5.9.6
pure-eval==0.2.2
py-cpuinfo==9.0.0
pycparser==2.21
pycryptodome==3.22.0
pydantic==2.11.5
pydantic_core==2.33.2
pydub==0.25.1
pyfluidsynth==1.3.4
pygame==2.6.1
Pygments==2.17.0
PyJWT==2.8.0
pyloudnorm==0.1.1
PyMatting==1.1.13
pynndescent==0.5.13
pyparsing==3.1.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
PyQt5-stubs==5.15.6.0
pyreadline3==3.5.4
pystoi==0.4.1
python-dateutil==2.8.2
python-engineio==4.12.0
python-json-logger==2.0.7
python-multipart==0.0.20
python-socketio==5.7.2
pytorch-wpe==0.0.1
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.1
randomname==0.2.1
referencing==0.31.0
regex==2024.11.6
rembg==2.0.65
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.13.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
ruff==0.11.7
safehttpx==0.1.6
safetensors==0.5.3
samplings==0.1.7
scikit-image==0.24.0
scikit-learn==1.3.2
scipy==1.12.0
seaborn==0.13.0
semantic-version==2.10.0
Send2Trash==1.8.2
sentencepiece==0.2.0
setuptools==68.2.2
shellingham==1.5.4
simple-websocket==1.1.0
simsimd==6.2.1
six==1.16.0
sniffio==1.3.0
sns==0.1
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.5
sox==1.5.0
soxr==0.5.0.post1
stack-data==0.6.3
starlette==0.46.2
statsmodels==0.14.4
stringzilla==3.12.5
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.18.0
tensorflow-aarch64==1.2
tensorflow_intel==2.18.0
termcolor==2.5.0
terminado==0.18.0
threadpoolctl==3.2.0
tifffile==2025.3.13
tiktoken==0.9.0
timm==1.0.15
tinycss2==1.2.1
tn==0.0.4
tokenizers==0.20.3
tomlkit==0.12.0
toolz==0.12.1
torch==2.5.1+cu118
torch-complex==0.4.4
torch-stoi==0.2.3
torchaudio==2.5.1+cu118
torchsde==0.2.6
torchvision==0.20.1
tornado==6.3.3
tqdm==4.66.2
traitlets==5.13.0
trampoline==0.1.2
transformers==4.46.2
typer==0.15.2
typer-config==1.4.0
types-python-dateutil==2.8.19.14
typing_extensions==4.13.2
typing-inspection==0.4.1
tzdata==2023.3
ultralytics==8.3.146
ultralytics-thop==2.0.14
umap-learn==0.5.7
uri-template==1.3.0
urllib3==2.0.7
uvicorn==0.29.0
watchdog==6.0.0
wcwidth==0.2.10
webcolors==1.13
webencodings==0.5.1
webrtcvad==2.0.10
websocket-client==1.6.4
websockets==11.0.3
Werkzeug==3.1.3
wheel==0.45.0
win32_setctime==1.2.0
wrapt==1.16.0
wsproto==1.2.0
yacs==0.1.8
yarl==1.18.3
zhipuai==1.0.7
zipp==3.21.0
